Poly2 Mixer and Mutator for use as RightFoldable for GeneticAlgorithm
